==> 3_SVM.ipynb <==
import pandas as pd
import numpy as np
import pickle
from sklearn.metrics import average_precision_score
from sklearn.model_selection import KFold, StratifiedKFold
from sklearn import svm
from functions import grupowanie, create_probs_series,stand

#wczytujemy wcześniej przygotowany zbiór

df=pd.read_excel("dumm_credit_card_data.xlsx",index_col="ID")

#zmienne objaśniające
features=pd.concat([df.loc[:,:"PAY_AMT6"],df.loc[:,"SEX_2":]],axis=1).columns.tolist()

def CV_SVM(df=df,n_splits=5,random_state=111,features=features,sample_size=3000,probability=True,*args,**kwargs):
    test_precisions=[]
    probs=[] 
    indices=[]
    skf=StratifiedKFold(n_splits=n_splits,shuffle=True,random_state=random_state)
    for train,test in skf.split(df[features],df.y):
        model=svm.SVC(probability=probability,*args,**kwargs,random_state=random_state)
        df2=df.iloc[train].sample(sample_size,random_state=random_state)
        model.fit(df2[features],df2.y)
        preds=model.predict_proba(df.iloc[test][features])[:,1].tolist()
        probs.append(preds)
        indices.append(test.tolist())
        test_precisions.append(average_precision_score(df.iloc[test].y,preds))
    return test_precisions, probs, indices

#sprawdźmy czy lepszy wynik uzyskamy używając standaryzacji na przedziale [0,1] czy z-score

features_st=['LIMIT_BAL',"AGE"]
features_st.extend(df.loc[:,"BILL_AMT1":"PAY_AMT6"].columns.tolist())

stand(features=features_st,df=df,valid=CV_SVM)

stand(features=features_st,df=df,valid=CV_SVM,function=lambda x: (x-np.mean(x))/np.std(x))

#lepszy wynik dla z-score
df[features_st]=df[features_st].apply(lambda x: (x-np.mean(x))/np.std(x))

#tuning hiperparametrów
results=[]
for kernel in ['linear','poly','rbf']:
    for C in [0.01,0.1,0.25,0.5,1,2,5,10,25,50,100]:
        if kernel=='poly':
            for degree in [2,3,4,5]:
                test_precisions, probs, indices=CV_SVM(degree=degree,C=C,kernel=kernel)
                print(f'C={C} kernel={kernel} degree={degree} precision=',np.mean(test_precisions),sep='')
                results.append((C,kernel,degree,np.mean(test_precisions)))
        elif kernel=='rbf':
            for gamma in [0.1,1,10,'scale']:
                test_precisions, probs, indices=CV_SVM(gamma=gamma,C=C,kernel=kernel)
                print(f'C={C} kernel={kernel} gamma={gamma} precision=',np.mean(test_precisions),sep='')
                results.append((C,kernel,gamma,np.mean(test_precisions)))
        else:
            #pomijamy C>2 dla kernela liniowego z uwagi na długi czas uczenia algorytmu
            if C>2:
                continue
            test_precisions, probs, indices=CV_SVM(C=C,kernel=kernel)
            print(f'C={C} kernel={kernel} precision=',np.mean(test_precisions),sep='')
            results.append((C,kernel,np.mean(test_precisions)))
                

#dla kernela liniowego i C większego od 5 podzieliłem próbki na mniejsze części z uwagi na bardzo długi czas uczenia algorytmu
for C in [5,10]:
    means=[]
    for k in range(3): 
        test_precisions, probs, indices=CV_SVM(C=C,kernel='linear',sample_size=1000)
        means.append(np.mean(test_precisions))          
    print(C,np.mean(means))
#widać, że wyniki dla tych C nie są lepsze niż w przypadku C<=2

results.sort(key=lambda x: x[-1],reverse=True)
results

#najlepszy model to C=10, kernel=rbf i gamma=scale
#sprobujmy dodac kwadrat zmiennej wiek

features2=features.copy()
features2.append("AGE_SQ")

df['AGE_SQ']=df[['AGE_SQ']].apply(lambda x: (x-np.mean(x))/np.std(x))

test_precisions, probs, indices=CV_SVM(features=features2,C=10)
precis=np.mean(test_precisions)
print(precis)
#pograsza model

#następnym krokiem będzie sprawdzenie, która zmienna prognozuje najlepiej
results=[]
for feature in features:
    test_precisions, probs, indices=CV_SVM(features=[feature],C=10)
    print(feature,np.mean(test_precisions))
    results.append((feature,np.mean(test_precisions)))

results.sort(key=lambda x: x[-1])
results

test_precisions, probs, indices=CV_SVM(C=10)
precis=np.mean(test_precisions)
print(precis)

#sprawdzmy jaki wynik da wyrzucanie najgorzej przewidujacych zmiennych
for result in results:
    features2=features.copy()
    features2.remove(result[0])
    test_precisions, probs, indices=CV_SVM(features=features2,C=10)
    precis2=np.mean(test_precisions)
    if precis2>precis:
        print(result[0],precis2)
        precis=precis2
        features=features2.copy()
    else:
        print('Brak poprawy')
        break

#identyczny wynik dla zmiennych education_4-6
#sprobuje polaczyc te zmienne
cols=df.loc[:,'EDUCATION_4':'EDUCATION_6'].columns.tolist()


grupowanie(df,columns=cols,new_column='EDUC4_6',valid=CV_SVM,function=sum,features=features,
          precision=precis,C=10)

test_precisions, probs, indices=CV_SVM(features=features,C=10)
precis=np.mean(test_precisions)
print(precis)

probs=create_probs_series(probs,indices)

modelSVM = {
    "nazwa":"SVM",
    "opis":"Model SVM, kernel rbf, C=10",
    "specyfikacja":{'kernel':"rbf", 'C':10, 'features':features.copy(),'df':df.copy()},
    "precyzja":precis,
    "predykcje":probs.copy(),
    }

#zapisujemy model

with open('ML_SVM.p','wb') as fp:
    pickle.dump(modelSVM,fp)
