==> 4_kNN.ipynb <==
import pandas as pd
import numpy as np
import pickle
from sklearn.metrics import average_precision_score
from sklearn.model_selection import KFold, StratifiedKFold
from functions import create_probs_series, stand

df=pd.read_excel("dumm_credit_card_data.xlsx",index_col="ID")

features=pd.concat([df.loc[:,:"PAY_AMT6"],df.loc[:,"SEX_2":]],axis=1).columns.tolist()

def CV_kNN(n_jobs=-1,df=df_kNN,n_splits=5,random_state=111,features=features_kNN,*args,**kwargs):
    test_precisions=[]
    probs=[]
    indices=[]
    skf=StratifiedKFold(n_splits=n_splits,shuffle=True,random_state=random_state)
    for train,test in skf.split(df[features],df.y):
        model=neighbors.KNeighborsClassifier(*args,**kwargs,n_jobs=n_jobs)
        model.fit(df.iloc[train][features],df.iloc[train].y)
        preds=model.predict_proba(df.iloc[test][features])[:,1].tolist()
        probs.append(preds)
        indices.append(test.tolist())
        test_precisions.append(average_precision_score(df.iloc[test].y,preds))
    return test_precisions, probs, indices

#sprawdźmy czy lepszy wynik uzyskamy używając standaryzacji na przedziale [0,1] czy z-score

features_st=['LIMIT_BAL',"AGE"]
features_st.extend(df.loc[:,"BILL_AMT1":"PAY_AMT6"].columns.tolist())

stand(features=features_st,df=df,valid=CV_kNN)

stand(features=features_st,df=df,valid=CV_kNN,function=lambda x: (x-np.mean(x))/np.std(x))

#lepszy wynik dla standaryzacji na przedziale [0,1]
df[features_st]=df[features_st].apply(lambda x: (x-np.min(x))/(np.max(x)-np.min(x)))

#tuning hiperparametrow
#optymalna liczba sasiadow
for k in [1, 3, 5, 10, 15, 30, 50, 100, 150, 200]:
    test_precisions, probs, indices=CV_kNN(n_neighbors=k)
    print(k,np.mean(test_precisions))

#sprawdzmy wieksza liczbe sasiadow
for k in [250,300,400,500]:
    test_precisions, probs, indices=CV_kNN(n_neighbors=k)
    print(k,np.mean(test_precisions))

#najlepszy wynik dla 250 sasiadow
#sprawdzmy czy nie lepszy wynik uzyskamy dla odleglosci miejskiej
test_precisions, probs, indices=CV_kNN(p=1,n_neighbors=250)
np.mean(test_precisions)
#pogorszenie

#sprobujmy dodac kwadrat zmiennej wiek

features2=features.copy()
features2.append("AGE_SQ")

df['AGE_SQ']=df[['AGE_SQ']].apply(lambda x: (x-np.min(x))/(np.max(x)-np.min(x)))

test_precisions, probs, indices=CV_kNN(n_neighbors=250,features=features2)
np.mean(test_precisions)

#brak poprawy

#sprawdzmy ktora zmienna najlepiej prognozuje
results=[]
for feature in features:
    test_precisions, probs, indices=CV_kNN(n_neighbors=250,features=[feature])
    print(feature,np.mean(test_precisions))
    results.append((feature,np.mean(test_precisions)))

results.sort(key=lambda x: x[1])
results

#sprobujmy przemnozyc najgorzej prognozujaca zmienna zmienna przez stala mniejsza od 1 
def feature_eng_kNN(consts=[0,0.1,0.25,0.5,0.75],feature_nr=0):
    for k in consts:
        df2=df.copy()
        df2[results[feature_nr][0]]*=k
        test_precisions, probs, indices=CV_kNN(df=df2,n_neighbors=250)
        print(k,np.mean(test_precisions))

feature_eng_kNN()

#brak poprawy
#przemnozmy teraz najlepsza zmienna przez stala wieksza od 1

feature_eng_kNN(consts=[1.2,1.5,1.75,2,2.5,3,3.5,4,5,6,7],feature_nr=-1)

#lekka poprawa
df[results[-1][0]]*=1.75

#najlepszy model knn
test_precisions, probs, indices=CV_kNN(df=df,features=features,n_neighbors=250)
precis=np.mean(test_precisions)
print(precis)

#Tworzymy serie pandasa zawierającą predykcje
probs=create_probs_series(probs,indices)
