==> 5_RF.ipynb <==
import pandas as pd
import numpy as np
import pickle
from sklearn.metrics import average_precision_score
from sklearn.model_selection import KFold, StratifiedKFold
from sklearn.ensemble import RandomForestClassifier
from functions import create_probs_series, grupowanie, log_func

#nastepny algorytm to Random Forest
#w przypadku algorytmu RF zmiennych kategorycznych nie zamieniamy na zerojedynkowe
#w przypadku naszego dfa nie ma potrzeby jednak uzywac label encodera, gdyz zmienne kategoryczne sa numeryczne
#nie trzeba tez standaryzowac zmiennych

df=pd.read_excel("credit_card_data.xlsx",index_col="ID")

features=df.loc[:,:"PAY_AMT6"].columns.tolist()

def CV_RF(n_splits=5,random_state=111,df=df,features=features,*args,**kwargs):
    test_precisions=[]
    probs=[]
    indices=[]
    skf=StratifiedKFold(n_splits=n_splits,shuffle=True,random_state=random_state)
    for train,test in skf.split(df[features],df.y):
        model=RandomForestClassifier(*args,**kwargs,random_state=random_state,n_jobs=-1)
        model.fit(df.iloc[train][features],df.iloc[train].y)
        preds=model.predict_proba(df.iloc[test][features])[:,1].tolist()
        probs.append(preds)
        indices.append(test.tolist())
        test_precisions.append(average_precision_score(df.iloc[test].y,preds))
    return test_precisions, probs, indices,model

test_precisions, probs, indices,model=CV_RF()
print(np.mean(test_precisions))

#tak jak w przypadku pozostalych algorytmow sprawdze czy dodanie kwadratu wieku poprawi model

features2=features.copy()
features2.append("AGE_SQ")

test_precisions, probs, indices,model=CV_RF(features=features2)
print(np.mean(test_precisions))

#przynioslo poprawe, teraz sprawdzmy czy dodanie kwadratu zmiennych PAY pomoze
features=features2.copy()

cols=df.loc[:,"PAY_0_SQ":"PAY_6_SQ"].columns.tolist()

features2.extend(cols)

test_precisions, probs, indices,model=CV_RF(features=features2)
print(np.mean(test_precisions))
#brak poprawy

#tuning hiperparametrów RF
#dla przyspieszenia obliczen do tuningu hiperparametrow uzylem 50 estymatorow
#optymalna maksymalna głębokość
for k in range(2,22,2):
    test_precisions, probs, indices,model=CV_RF(features=features,n_estimators=50,max_depth=k)
    print(k,np.mean(test_precisions))

#optymalna maksymalna ilosc zmiennych uzywana przy trenowaniu algorytmu
for k in range(3,11):
    test_precisions, probs, indices,model=CV_RF(features=features,n_estimators=50,max_depth=10,max_features=k)
    print(k,np.mean(test_precisions))

#minimalna ilosc obserwacji przed podzialem
for k in [2, 4, 6, 8, 10, 15, 20, 30]:
    test_precisions, probs, indices,model=CV_RF(features=features,n_estimators=50,max_depth=10,max_features=8,min_samples_split=k)
    print(k,np.mean(test_precisions))

#minimalna ilosc obserwacji po podziale
for k in [1,2,3,4,5]:
    test_precisions, probs, indices,model=CV_RF(features=features,n_estimators=50,max_depth=10,max_features=8,min_samples_split=10,min_samples_leaf=k)
    print(k,np.mean(test_precisions))

#liczba drzew
for k in [50,75,100,125,150,175,200]:
    test_precisions, probs, indices,model=CV_RF(features=features,n_estimators=k,max_depth=10,max_features=8,min_samples_split=10,min_samples_leaf=1)   
    print(k,np.mean(test_precisions))

test_precisions, probs, indices,model=CV_RF(features=features,n_estimators=100,max_depth=10,max_features=8,min_samples_split=10,min_samples_leaf=1)   
precis=np.mean(test_precisions)
print(precis)

#sprawdzmy udzial poszczegolnych zmiennych w prognozowaniu
importances=list(zip(features,model.feature_importances_.tolist()))
importances.sort(key= lambda x: x[1],reverse=True)
importances

#najlepsze argumenty z grid searcha

parametersRF={'n_estimators':100,'max_depth':10,'max_features':8,'min_samples_split':10,'min_samples_leaf':1}

parametersRF

#grupowanie malo istotnych zmiennych o podobnych wspolczynnikach

grupowanie(df=df,columns=['BILL_AMT4','BILL_AMT5'],new_column='BILL_AMT45',
          valid=CV_RF,function='mean',features=features,precision=precis,**parametersRF)

test_precisions, probs, indices,model=CV_RF(df=df,features=features,**parametersRF)   
precis=np.mean(test_precisions)

grupowanie(df=df,columns=['BILL_AMT3','BILL_AMT6'],new_column='BILL_AMT36',
          valid=CV_RF,function='mean',features=features,precision=precis,**parametersRF)

test_precisions, probs, indices,model=CV_RF(df=df,features=features,**parametersRF)   
precis=np.mean(test_precisions)

grupowanie(df=df,columns=['PAY_AMT5','PAY_AMT6'],new_column='PAY_AMT56',
          valid=CV_RF,function='mean',features=features,precision=precis,**parametersRF)

#kolejne grupowanie nie ulepsza modelu
log_func(df,'PAY_AMT2',valid=CV_RF,features=features,precision=precis,**parametersRF)
#tak samo logarytmowanie skosnej zmiennej

cols=df.loc[:,'PAY_0':'PAY_6'].columns.tolist()
df['PAY_max']=df[cols].apply('max',axis=1)

features2=features.copy()
features2.append("PAY_max")

test_precisions, probs, indices,model=CV_RF(features=features2,df=df,**parametersRF)   
print(np.mean(test_precisions))
#poprawa po dodaniu maksymalnego zwlekania z zaplata

features=features2.copy()

test_precisions, probs, indices,model=CV_RF(features=features,df=df,**parametersRF)   
print(np.mean(test_precisions))

#warto poszukac optymalnych hiperparametrow za pomoca random search

results=[]
for k in range(50):
    parameters={'n_estimators':np.random.choice(range(75,180,5)),
                'max_depth':np.random.randint(8,14),
                'max_features':np.random.randint(4,11),
                'min_samples_split':np.random.randint(2,11)
               }
    parameters['min_samples_leaf']=np.random.randint(1,parameters['min_samples_split']//2+1)
    test_precisions, probs, indices,model=CV_RF(**parameters,features=features,df=df)   
    results.append((parameters.copy(),np.mean(test_precisions)))
    print(np.mean(test_precisions))

sorted(results,key=lambda x: x[1],reverse=True)[0]
#brak poprawy

test_precisions, probs, indices,model=CV_RF(features=features,df=df,**parametersRF)   
precis=np.mean(test_precisions)
print(precis)

probs=create_probs_series(probs,indices)

parametersRF

modelRF = {
    "nazwa":"RF",
    "opis":"Model RF, max_depth=10, max_features=8, min_samples_split=10, min_samples_leaf=1",
    "specyfikacja":{'n_estimators':100,'max_depth':10, 'max_features':8, 'min_samples_split':10,
                    'min_samples_leaf':1,'features':features,'df':df},     
    "precyzja":precis,
    "predykcje":probs.copy(),
    }

with open('ML_RF.p','wb') as fp:
    pickle.dump(modelRF,fp)
