==> 6_XGB.ipynb <==
import pandas as pd
import numpy as np
import pickle
from sklearn.metrics import average_precision_score
from sklearn.model_selection import KFold, StratifiedKFold
import xgboost as xgb
from functions import create_probs_series
import random

#boostowane drzewa decyzyjne

df=pd.read_excel("credit_card_data.xlsx",index_col="ID")

features=df.loc[:,:"PAY_AMT6"].columns.tolist()

#tworzymy funkcje do wywolywania boostowanych drzew decyzyjnych

def run_xgb(train,validate,features,target,
           eta=0.03,max_depth=7,num_boost_round=1000,early_stopping_rounds=50,
           lambdaX=0,alpha=0,gamma=0,min_child_weight=0,
           subsample=0.7, colsample_bytree=0.7, colsample_bylevel=1,
           tree_method='hist',grow_policy='depthwise',eval_metric='aucpr', seed=111,
           objective='binary:logistic', verbosity=0, booster='gbtree',verbose_eval=False):
    dtrain=xgb.DMatrix(train[features],train[target],feature_names=features)
    dvalidate=xgb.DMatrix(validate[features],validate[target],feature_names=features)
    evals=[(dtrain,'train'),(dvalidate,'validate')]
    params=['eta','max_depth','alpha','gamma','min_child_weight','subsample','colsample_bytree',
           'colsample_bylevel','tree_method','grow_policy','eval_metric','seed','objective','verbosity',
           'booster']
    Params=dict()
    for param in params:
        Params[param]=eval(param)
    Params['lambda']=lambdaX
    clf=xgb.train(Params,dtrain,num_boost_round,evals,early_stopping_rounds=early_stopping_rounds,verbose_eval=verbose_eval)
    score=clf.best_score
    valResults=clf.predict(dvalidate,ntree_limit=clf.best_ntree_limit)
    imp=clf.get_fscore()
    imp=[(x,y) for x,y in imp.items()]
    imp.sort(key=lambda x: x[1],reverse=True)
    return score,valResults,imp
    

#na podstawie poprzedniej funkcji tworzymy wrapper do walidacji krzyzowej

def CV_XGB(df=df,n_splits=5,random_state=111,features=features,*args,**kwargs):
    test_precisions=[]
    probs=[]
    indices=[]
    skf=StratifiedKFold(n_splits=n_splits,shuffle=True,random_state=random_state)
    for train,test in skf.split(df[features],df.y):
        score,valResults,imp=run_xgb(df.iloc[train],df.iloc[test],features,'y',seed=random_state,*args,**kwargs)
        probs.append(valResults.tolist())
        indices.append(test.tolist())
        test_precisions.append(score)
    return test_precisions, probs, indices, imp

test_precisions, probs, indices, imp=CV_XGB()
print(np.mean(test_precisions))

#tak jak w przypadku pozostalych algorytmow sprawdze czy dodanie kwadratu wieku poprawi model

features2=features.copy()
features2.append("AGE_SQ")

test_precisions, probs, indices, imp=CV_XGB(features=features2)
print(np.mean(test_precisions))

#nie poprawilo modelu

#sprawdzmy czy dodanie kwadratu zmiennych PAY pomoze

features2=features.copy()

cols=df.loc[:,"PAY_0_SQ":"PAY_6_SQ"].columns.tolist()
features2.extend(cols)

test_precisions, probs, indices, imp=CV_XGB(features=features2)
print(np.mean(test_precisions))

#brak poprawy

#sprawdzmy jak bardzo przetrenowuje sie algorytm

test_precisions, probs, indices, imp=CV_XGB(features=features,verbose_eval=True)
print(np.mean(test_precisions))

#algorym nie przetrenowuje sie w znaczacym stopniu, wiec tuning hiperparametrow zaczne od parametrow innych niz regularyzacyjne    

#najpierw sprawdze glebokosc drzewa
for k in range(4,16,2):
    test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=k)
    print(k,np.mean(test_precisions))

# % uzytych obserwacji
for k in range(2,10):
    test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=6,subsample=k/10)
    print(k/10,np.mean(test_precisions))

# % zmiennych uzytych na drzewo
for k in range(2,10):
    test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=6,subsample=0.9,colsample_bytree=k/10)
    print(k/10,np.mean(test_precisions))

# % zmiennych uzytych na poziom
for k in range(2,11):
    test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=6,subsample=0.9,colsample_bytree=0.7,colsample_bylevel=k/10)
    print(k/10,np.mean(test_precisions))

#regularyzacja l2
for k in range(8):
    test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=6,subsample=0.9,colsample_bytree=0.7,colsample_bylevel=1,lambdaX=k**2)
    print(k**2,np.mean(test_precisions))

#gamma 
for k in [0,0.1,0.25,0.5,1,2,5,6,8,10]:
    test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=6,subsample=0.9,colsample_bytree=0.7,colsample_bylevel=1,lambdaX=16,gamma=k)
    print(k,np.mean(test_precisions))

#regularyzacja l1
for k in range(8):
    test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=6,subsample=0.9,colsample_bytree=0.7,colsample_bylevel=1,lambdaX=16,gamma=0,alpha=k**2)
    print(k**2,np.mean(test_precisions))

#learning rate
for k in range(1,11):
     test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=6,subsample=0.9,colsample_bytree=0.7,colsample_bylevel=1,lambdaX=16,gamma=0,alpha=0,eta=k/100)
     print(k/100,np.mean(test_precisions))

test_precisions, probs, indices, imp=CV_XGB(features=features,max_depth=6,subsample=0.9,colsample_bytree=0.7,colsample_bylevel=1,lambdaX=16,gamma=0,alpha=0,eta=0.03)      
precis=np.mean(test_precisions)
print(precis)

#najwazniejsze zmienne w XGB

imp

#najczesciej uzywany LIMIT_BAL, sprobujmy dodac kwadrat tej zmiennej

df['LIMIT_BAL_SQ']=df['LIMIT_BAL']**2
features2=features.copy()
features2.append('LIMIT_BAL_SQ')
test_precisions, probs, indices, imp=CV_XGB(features=features2,max_depth=6,subsample=0.9,colsample_bytree=0.7,colsample_bylevel=1,lambdaX=16,gamma=0,alpha=0,eta=0.03)    
precis=np.mean(test_precisions)
print(precis)

#brak poprawy
#sprobujmy zmienic zmienne kategoryczne za pomoca one hot encodingu

df2=df.copy()
df2.loc[:,'SEX':'MARRIAGE']=df2.loc[:,'SEX':'MARRIAGE'].astype('object')
df2=pd.get_dummies(df2,drop_first=True)

features2=features.copy()
features2.remove("MARRIAGE")
features2.remove("SEX")
features2.remove("EDUCATION")
features2.extend(df2.loc[:,"SEX_2":].columns.tolist())

test_precisions, probs, indices, imp=CV_XGB(features=features2,df=df2,max_depth=6,subsample=0.9,colsample_bytree=0.7,colsample_bylevel=1,lambdaX=16,gamma=0,alpha=0,eta=0.03)    
precis=np.mean(test_precisions)
print(precis)
#pogarsza model

#sprobujmy przeprowadzic jeszcze raz tuning hiperparametrow za pomoca random search

results=[]
for k in range(50):
    parameters={'max_depth':random.randint(6,13),
               'subsample':random.uniform(0.55,0.95),
               'colsample_bytree':random.uniform(0.6,1),
               'colsample_bylevel':random.uniform(0.6,1),
               'lambdaX':random.randint(0,20),
               'alpha':random.uniform(0,2),
               'gamma':random.choice([0,0.01,0.05,0.1,0.25,0.5,1,2,3,4,5]),
               'eta':random.choice([0.025,0.03,0.04,0.045])
               }
    test_precisions, probs, indices, imp=CV_XGB(features=features,**parameters)
    results.append((parameters.copy(),np.mean(test_precisions)))
    print(np.mean(test_precisions))

sorted(results,key=lambda x: x[1],reverse=True)

#poprawa wyniku

test_precisions, probs, indices, imp=CV_XGB(features=features,lambdaX=13,gamma=0,alpha=1.69,max_depth=8,subsample=0.565,colsample_bytree=0.69,colsample_bylevel=0.96,eta=0.025)
precis=np.mean(test_precisions)
print(precis)

probs=create_probs_series(probs,indices)

modelXGB = {
    "nazwa":"XGB",
    "opis":"Model XGB, max_depth=8, subsample=0.565, colsample_bytree=0.69, colsample_bylevel=0.96, lambda=13, gamma=0, alpha=1.69",
    "specyfikacja":{'eta':0.025,'max_depth':8, 'subsample':0.565, 'colsample_bytree':0.69,
                    'colsample_bylevel':0.96,'lambdaX':13,'gamma':0,'alpha':1.69,'features':features,'df':df},     
    "precyzja":precis,
    "predykcje":probs.copy(),
    }

with open('ML_XGB.p','wb') as fp:
    pickle.dump(modelXGB,fp)
